+++ src/backend.rs
@@ -28,7 +28,7 @@ use crate::error::{Error, MeasureError, SetCursorError};
 
 pub use crate::wrapper::Wrapper;
 #[cfg(feature = "alloc")]
-pub use crate::wrapper::blink::BlinkWrapper;
+pub use crate::wrapper::blink::{BlinkWrapper, ConfigureBlinkWrapper};
 pub use crate::wrapper::flush::FlushWrapper;
 
 /// Backend for Ratatui that renders to a display with the [`embedded-graphics`](embedded_graphics)